Moved the JSON constants to a new interface
There are needed on replayer side to unparse the JSON. So far,
they are hard wired, which is bad. With this, we can import these
constants and not need to import the whole ClientRecorder class
just for them.

These might be moved to a special "util" or "common" project if
there will ever be a need.