COPELogger is a singleton, RecorderFacade is not
So I am using each one in their new specific way.

The reason I chose this, is that we might need to have multiple
recorders in one JVM (IntelliJ thing). This logic could also be applied
to the logger, but it's quite difficult to make it not a singleton
right now.

Ideally, there should not be a single singleton left!